docs(module-resolver): test against error code instead of instances
