+++ packages/@lwc/module-resolver/src/__tests__/module-errors.spec.ts
@@ -6,17 +6,15 @@
  */
 
 import { resolveModule } from '../index';
-import { LwcConfigError } from '../errors';
-
-import { fixture } from './test-utils';
+import { fixture, NO_LWC_MODULE_FOUND_CODE, LWC_CONFIG_ERROR_CODE } from './test-utils';
 
 describe('resolution errors', () => {
     test('throw when no lwc config is present in the path', () => {
         const specifier = 'missing';
         const importer = '/errors/non/existent/module.js';
 
-        expect(() => resolveModule(specifier, importer)).toThrowErrorWithType(
-            LwcConfigError,
+        expect(() => resolveModule(specifier, importer)).toThrowErrorWithCode(
+            LWC_CONFIG_ERROR_CODE,
             `Invalid LWC configuration in "${importer}". Unable to find any LWC configuration file.`
         );
     });
@@ -25,8 +23,8 @@ describe('resolution errors', () => {
         const specifier = 'missing';
         const importer = fixture('errors/invalid-lwc-config/invalid-lwc-config.js');
 
-        expect(() => resolveModule(specifier, importer)).toThrowErrorWithType(
-            LwcConfigError,
+        expect(() => resolveModule(specifier, importer)).toThrowErrorWithCode(
+            LWC_CONFIG_ERROR_CODE,
             `Invalid LWC configuration in "${fixture(
                 'errors/invalid-lwc-config'
             )}". "lwc.config.json" must be at the package root level along with the "package.json".`
@@ -37,8 +35,8 @@ describe('resolution errors', () => {
         const specifier = 'something';
         const importer = fixture('errors/invalid-record/invalid-record.js');
 
-        expect(() => resolveModule(specifier, importer)).toThrowErrorWithType(
-            LwcConfigError,
+        expect(() => resolveModule(specifier, importer)).toThrowErrorWithCode(
+            LWC_CONFIG_ERROR_CODE,
             `Invalid LWC configuration in "${fixture(
                 'errors/invalid-record'
             )}". Unknown module record "{}"`
@@ -49,8 +47,8 @@ describe('resolution errors', () => {
         const specifier = 'something';
         const importer = fixture('errors/invalid-alias/invalid-alias.js');
 
-        expect(() => resolveModule(specifier, importer)).toThrowErrorWithType(
-            LwcConfigError,
+        expect(() => resolveModule(specifier, importer)).toThrowErrorWithCode(
+            LWC_CONFIG_ERROR_CODE,
             `Invalid LWC configuration in "${fixture(
                 'errors/invalid-alias'
             )}". Unknown module record "{"name":"something"}"`
@@ -61,7 +59,8 @@ describe('resolution errors', () => {
         const specifier = 'empty';
         const importer = fixture('errors/empty/empty.js');
 
-        expect(() => resolveModule(specifier, importer)).toThrow(
+        expect(() => resolveModule(specifier, importer)).toThrowErrorWithCode(
+            NO_LWC_MODULE_FOUND_CODE,
             `Unable to resolve "${specifier}" from "${importer}"`
         );
     });
